Fixed an issue where the ball material being referenced was the raw copy on desk rather than an instance
It turns out that if you have a child two layers deep within the
hierarchy of a prefab that has a material attached to a mesh renderer,
that material will be the raw copy on disk rather than an instance. This
leads to all kinds of issues and I have no idea why Unity does this.
Regardless, turning them into separate prefabs and then parenting them
via a script fixes the problem.